update DB & EF

+++ Tables/XpImportLog.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 
-#nullable disable
-
 namespace HrAdm.Tables
 {
     public partial class XpImportLog
@@ -10,9 +8,9 @@ namespace HrAdm.Tables
         public string Id { get; set; }
         public string Type { get; set; }
         public string FileName { get; set; }
-        public int OkCount { get; set; }
-        public int FailCount { get; set; }
-        public int TotalCount { get; set; }
+        public short OkCount { get; set; }
+        public short FailCount { get; set; }
+        public short TotalCount { get; set; }
         public string CreatorName { get; set; }
         public DateTime Created { get; set; }
     }
